[3.x] Add Template Types
The goal of this change set is to add template types to the cache package, and thus make all use of it type safe. As we transition ReactPHP into fully typed in v3 we can't ignore doing this with the cache package. But forcing our uses to supply a type, we can ensure through our entire stack we get what we expect. For example DNS would use this to type it's cache to `\React\Dns\Model\Message` as shown in this change set: https://github.com/reactphp/dns/compare/3.x...WyriHaximus-labs:dns:3.x-typed-cache